Improve stage 3 speed and success rate dramatically (#24)
* Improve stage 3 speed and success rate dramatically

* Move CIPHER_TEXT_DATA structure to a cachable heap allocation

This on its own, might affect hit rate compared to the previous stack allocation,
but it will ensure we will not overflow if the payload is ran on a
smaller stack.

* Use assembler code for ciphertext overwriter loop

This ensures the stack is not touched in this critical loop, meaning the
only memory that actually has to stay in cache is the lookup table and
it no longer matters if it is on the stack or elsewhere.

Assembler is so much faster than optimized C here that it shifts the
timing of hits noticably towards earlier blocks. To compensate for this,
I've added a "nop loop" between the detection of a matching canary and
the start of doing the overwrites.

This assembler block is based on:
https://github.com/grimdoomer/Xbox360BadUpdate/pull/24#discussion_r2252667295